requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
colorama==0.4.6
distro==1.9.0
Django==5.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
openai==1.14.2
pillow==10.2.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1